mise: Use in ci-mgmt's update workflow (#1911)
Simplify (and speed up) the setup of our update-workflows jobs. The only
tools we need to install are the provider's version of go and
golangci-lint -- and we can cache those.

Confirmed working
[here](https://github.com/pulumi/ci-mgmt/actions/runs/20113027371/job/57715271370).

As a reminder, eventually we want to get rid of this in favor of
something providers invoke, instead of ci-mgmt invoking whatever
workflows the provider happens to have committed.