Update workflows using GitHub app auth (#1913)
I've created a GitHub app (Pulumi Provider Automation) and added its
ID/key to the `pulumi/imports/github-secrets` environment.

This PR starts using it in ci-mgmt's workflows where we saw some
throttling of the pulumi-bot user.

I'm also adding plumbing so we can hook this up in provider jobs as
well. If pulumi-bot is bumping into rate limits then this is something
we'll eventually need to do anyway.

When we do start using this in other places, one thing to watch out for
is a lot of our automation special-cases the pulumi-bot user
specifically. With this app token the user will be different, so we need
to account for that.

Fixes https://github.com/pulumi/ci-mgmt/issues/1910.