Flag upgrade-provider build failures (#1927)
We have a visibility gap right now around upgrade-provider failures.
These failures get flagged in issues like [this
one](https://github.com/pulumi/pulumi-azure/issues/3367), but not in a
way that surfaces them to our dashboards. For example I can't tell right
now if this workflow is broken for _all_ providers or just azure.

This change causes the upgrade-provider workflow to fail if we weren't
able to create a PR. This will get the failure picked up by the same
automation we use for all our other workflow failures: an issue will be
created and visible on our dashboards, and the issue will auto-resolve
when the problem is fixed.